refactor: stylelint 재적용 및 중복코드 컴포넌트 사용으로 간소화 진행
