+++ src/assets/svg/index.ts
@@ -1,14 +1,13 @@
 import Icn_arrow_narrow_right from '@/assets/svg/arrow-narrow-right.svg?react';
+import Icn_calander from '@/assets/svg/calendar-minus-01.svg?react';
+import Icn_date_clock from '@/assets/svg/clock.svg?react';
 import Icn_clock from '@/assets/svg/icn_clock.svg?react';
 import Icn_nav_calendar from '@/assets/svg/icn_nav_calendar.svg?react';
 import Icn_nav_dashboard from '@/assets/svg/icn_nav_dashboard.svg?react';
 import Icn_nav_setting from '@/assets/svg/icn_nav_setting.svg?react';
 import Icn_nav_today from '@/assets/svg/icn_nav_today.svg?react';
-
-import Icn_calander from '../../assets/svg/calendar-minus-01.svg?react';
-import Icn_date_clock from '../../assets/svg/clock.svg?react';
-import Icn_line from '../../assets/svg/line164.svg?react';
-import IcnXCricle from '../../assets/svg/x-circle.svg?react';
+import Icn_line from '@/assets/svg/line164.svg?react';
+import IcnXCricle from '@/assets/svg/x-circle.svg?react';
 
 const Icons = {
 	Icn_clock,
