+++ src/components/common/BtnDate/BtnDate.tsx
@@ -1,5 +1,8 @@
+import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 import { useState } from 'react';
+
+import BtnDateText, { TextWrapper } from './BtnDateText';
 
 import Icons from '@/assets/svg/index';
 
@@ -9,7 +12,7 @@ interface BtnDateProps {
 	size?: string;
 }
 
-const BtnDate = (props: BtnDateProps) => {
+function BtnDate(props: BtnDateProps) {
 	const { date = '마감 기한', time = '마감 시간', size = 'big' } = props;
 	const [isPressed, setIsPressed] = useState(false);
 	const [isClicked, setIsClicked] = useState(false);
@@ -36,23 +39,13 @@ const BtnDate = (props: BtnDateProps) => {
 			onMouseDown={handleMouseDown}
 			onMouseUp={handleMouseUp}
 		>
-			<BtnDateContainer>
-				<CalanderIcon />
-				<TextWrapper isDefault={isDefaultDate} size={size}>
-					{date}
-				</TextWrapper>
-			</BtnDateContainer>
+			<BtnDateText icon={<CalanderIcon />} text={date} isDefault={isDefaultDate} size={size} />
 			<LineIcon size={size} />
-			<BtnDateContainer>
-				<ClockIcon />
-				<TextWrapper isDefault={isDefaultTime} size={size}>
-					{time}
-				</TextWrapper>
-			</BtnDateContainer>
+			<BtnDateText icon={<ClockIcon />} text={time} isDefault={isDefaultTime} size={size} />
 			<XIcon isClicked={isClicked} />
 		</BtnDateLayout>
 	);
-};
+}
 
 export default BtnDate;
 
@@ -60,48 +53,33 @@ const XIcon = styled((props: React.SVGProps<SVGSVGElement> & { isClicked: boolea
 	const { isClicked, ...rest } = props;
 	return <Icons.IcnXCricle {...rest} />;
 })<{ isClicked: boolean }>`
+	display: ${({ isClicked }) => (isClicked ? 'flex' : 'none')};
 	width: 2rem;
 	height: 2rem;
-	display: ${({ isClicked }) => (isClicked ? 'flex' : 'none')};
 `;
 
 const CalanderIcon = styled(Icons.Icn_calander, { target: 'CalanderIcon' })`
 	display: flex;
+	align-items: center;
+	justify-content: center;
 	width: 1.4rem;
 	height: 1.4rem;
-	justify-content: center;
-	align-items: center;
 `;
 
 const ClockIcon = styled(Icons.Icn_date_clock, { target: 'ClockIcon' })`
 	display: flex;
+	align-items: center;
+	justify-content: center;
 	width: 1.4rem;
 	height: 1.4rem;
-	justify-content: center;
-	align-items: center;
 `;
 
 const LineIcon = styled(Icons.Icn_line, { target: 'LineIcon' })<{ size: string }>`
 	display: flex;
+	align-items: center;
+	justify-content: center;
 	width: 0.1rem;
 	height: ${({ size }) => (size === 'big' ? '2.2rem' : '1.2rem')};
-	justify-content: center;
-	align-items: center;
-`;
-
-const TextWrapper = styled('div', { target: 'TextWrapper' })<{ isDefault: boolean; size: string }>`
-	display: flex;
-	flex-direction: column;
-	align-items: flex-start;
-	gap: 1rem;
-	color: ${({ isDefault, theme }) => (isDefault ? theme.palette.grey5 : theme.palette.black)};
-	font-size: ${({ size }) => (size === 'big' ? '1.4rem' : '1.2rem')};
-`;
-
-const BtnDateContainer = styled.div`
-	display: flex;
-	align-items: center;
-	gap: 0.4rem;
 `;
 
 const BtnDateLayout = styled.div<{
@@ -112,35 +90,42 @@ const BtnDateLayout = styled.div<{
 	isDefaultTime: boolean;
 }>`
 	display: flex;
+	gap: 1rem;
+	align-items: center;
 	width: fit-content;
 	min-width: 1.8rem;
 	padding: 0.5rem 1rem;
-	gap: 1rem;
-	align-items: center;
-	border-radius: 0.8rem;
-	border: 0.1rem solid ${({ theme }) => theme.palette.grey3};
+
 	background: ${({ theme }) => theme.palette.white};
 	cursor: pointer;
+	border: 1px solid ${({ theme }) => theme.palette.grey3};
 	border-color: ${({ isClicked, theme }) => (isClicked ? theme.palette.primary : theme.palette.grey3)};
-
-	${({ isPressed }) =>
-		isPressed &&
-		`
-		border-width: 0rem;
-	`}
+	border-radius: 8px;
 
 	${({ isClicked, size, theme }) =>
 		isClicked &&
-		`
-		padding-right: ${size === 'big' ? '0.6rem' : '0.2rem'};
-		border-width: 0.2rem;
-		border-color: ${theme.palette.primary};
-	`}
+		css`
+			padding-right: ${size === 'big' ? '0.6rem' : '0.2rem'};
+
+			border-color: ${theme.palette.primary};
+			border-width: 2px;
+		`}
+
+	${({ isPressed, theme }) =>
+		isPressed &&
+		css`
+			border-width: 0;
+
+			${TextWrapper} {
+				color: ${theme.palette.white};
+			}
+		`}
 
 	&:hover {
-		background: ${({ isPressed, theme }) => (isPressed ? theme.palette.grey5 : theme.palette.grey4)};
 		color: ${({ isPressed, theme }) => (isPressed ? theme.palette.white : theme.palette.grey6)};
-		border-width: 0rem;
+
+		background: ${({ isPressed, theme }) => (isPressed ? theme.palette.grey5 : theme.palette.grey4)};
+		border-width: 0;
 
 		${TextWrapper} {
 			color: ${({ isDefaultDate, isDefaultTime, theme }) =>
@@ -148,18 +133,11 @@ const BtnDateLayout = styled.div<{
 		}
 	}
 
-	${({ isPressed, theme }) =>
-		isPressed &&
-		`
-		${TextWrapper} {
-			color: ${theme.palette.white};
-		}
-	`}
-
 	&:hover ${CalanderIcon}, &:hover ${ClockIcon}, &:hover ${LineIcon} {
 		path {
 			stroke: ${({ isPressed, theme }) => (isPressed ? theme.palette.grey4 : theme.palette.grey5)};
 		}
+
 		line {
 			stroke: ${({ isPressed, theme }) => (isPressed ? theme.palette.grey4 : theme.palette.grey5)};
 		}
