+++ src/pages/Today.tsx
@@ -1,3 +1,5 @@
+import BtnDate from '@/components/common/BtnDate/BtnDate';
+import BtnStagingDate from '@/components/common/BtnDate/BtnStagingDate';
 import NavBar from '@/components/common/NavBar';
 import TextboxDailydate from '@/components/common/textbox/TextboxDailydate';
 import TextboxInput from '@/components/common/textbox/TextboxInput';
@@ -5,8 +7,6 @@ import TextInputDesc from '@/components/common/textbox/TextInputDesc';
 import TextInputStaging from '@/components/common/textbox/TextInputStaging';
 import TextInputTime from '@/components/common/textbox/TextInputTime';
 import TextInputTitle from '@/components/common/textbox/TextInputTitle';
-import BtnDate from '@/components/commons/BtnDate';
-import BtnStagingDate from '@/components/commons/BtnStagingDate';
 
 function Today() {
 	return (
